More Optimizations for New State (#4641)
* map optimization

* more optimizations

* use a custom hasher

* comment

* block operations optimizations

* Update beacon-chain/state/types.go


* fixed up various operations to use the validator index map access

+++ beacon-chain/blockchain/forkchoice/metrics.go
@@ -85,7 +85,7 @@ func reportEpochMetrics(state *stateTrie.BeaconState) {
 
 	validators := state.Validators()
 	for i, validator := range validators {
-		valBalance, err := state.BalanceAtIndex(i)
+		valBalance, err := state.BalanceAtIndex(uint64(i))
 		if err != nil {
 			log.WithError(err).Error("could not get balance for validator")
 			return
